refactor: implementa react query na listagem de tarefas
